% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{plotFacialKeypoints}
\alias{plotFacialKeypoints}
\title{Plot some images incl. keypoints}
\usage{
plotFacialKeypoints(imgSet, imgIndex, keypointPositions = NULL,
  meanIntensity = NULL, histEqualize = FALSE, order = c("normalize",
  "equalize"))
}
\arguments{
\item{imgSet}{The image set as Matrix with 9216 columns}

\item{imgIndex}{The index of the image that should be plotted}

\item{keypointPositions}{The keypoint positions as data.frame}

\item{meanIntensity}{Value between 0 and 255 determining the average intensity the printed image should have}

\item{histEqualize}{If set to true the image will be histogram equalized before printing}

\item{order}{Determines wether the image should be normalized or equalized first if both functions should be applied}
}
\examples{
plotFacialKeypoints(im.train, 1, d.train)
plotFacialKeypoints(im.test, 1)

}

