% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchSearch.R
\name{predictLearner.regr.patchSearch}
\alias{predictLearner.regr.patchSearch}
\title{Title}
\usage{
predictLearner.regr.patchSearch(.learner, .model, .newdata, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

